requirements.txt
django>=4.0.0,<4.1.0
djangorestframework
pyyaml
requests
django-cors-headers
psycopg2
celery[redis]
flower
django-phonenumber-field[phonenumbers]
django-sms
djangorestframework-simplejwt[crypto]
